Removal of openssl_crypto.c
Splits functions from openssl_crypto.c into many small files
