+++ src/lib/crypto.c
@@ -87,7 +87,7 @@ __RCSID("$NetBSD: crypto.c,v 1.36 2014/02/17 07:39:19 agc Exp $");
 \return length of MPI
 \note only RSA at present
 */
-int 
+int
 pgp_decrypt_decode_mpi(uint8_t *buf,
 				unsigned buflen,
 				const BIGNUM *g_to_k,
@@ -198,7 +198,7 @@ pgp_decrypt_decode_mpi(uint8_t *buf,
 \ingroup Core_MPI
 \brief RSA-encrypt an MPI
 */
-unsigned 
+unsigned
 pgp_rsa_encrypt_mpi(const uint8_t *encoded_m_buf,
 		    const size_t sz_encoded_m_buf,
 		    const pgp_pubkey_t * pubkey,
@@ -235,7 +235,7 @@ pgp_rsa_encrypt_mpi(const uint8_t *encoded_m_buf,
 \ingroup Core_MPI
 \brief Elgamal-encrypt an MPI
 */
-unsigned 
+unsigned
 pgp_elgamal_encrypt_mpi(const uint8_t *encoded_m_buf,
 		    const size_t sz_encoded_m_buf,
 		    const pgp_pubkey_t * pubkey,
@@ -349,7 +349,7 @@ Encrypt a file
 \param allow_overwrite Allow output file to be overwrwritten if it exists
 \return 1 if OK; else 0
 */
-unsigned 
+unsigned
 pgp_encrypt_file(pgp_io_t *io,
 			const char *infile,
 			const char *outfile,
@@ -444,7 +444,7 @@ pgp_encrypt_buf(pgp_io_t *io,
    \param getpassfunc Callback to use to get passphrase
 */
 
-unsigned 
+unsigned
 pgp_decrypt_file(pgp_io_t *io,
 			const char *infile,
 			const char *outfile,
@@ -615,4 +615,18 @@ pgp_decrypt_buf(pgp_io_t *io,
 
 	/* if we didn't get the passphrase, return NULL */
 	return (parse->cbinfo.gotpass) ? outmem : NULL;
+}
+
+void pgp_crypto_finish(void)
+{
+	// currently empty implementation
+}
+
+BIGNUM* new_BN_take_mp(botan_mp_t mp)
+{
+   PGPV_BIGNUM  *a;
+
+   a = calloc(1, sizeof(*a));
+   a->mp = mp;
+   return a;
 }
